+++ frontend/src/fallback/common/components/organisms/AnnouncementWrapper/AnnouncementWrapper.tsx
@@ -1,0 +1,5 @@
+const AnnouncementWrapper = () => {
+  return <></>;
+}
+
+export default AnnouncementWrapper;
